requirements.txt
cryptography
PyMySQL
python-dotenv
pydantic
rich
python-docx
